// frontend/src/generated/models/JsonWebSignatureAlgorithm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Small KMS Admin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const JsonWebSignatureAlgorithm = {
    Es256: 'ES256',
    Es384: 'ES384',
    Es512: 'ES512',
    Ps256: 'PS256',
    Ps384: 'PS384',
    Ps512: 'PS512',
    Rs256: 'RS256',
    Rs384: 'RS384',
    Rs512: 'RS512'
} as const;
export type JsonWebSignatureAlgorithm = typeof JsonWebSignatureAlgorithm[keyof typeof JsonWebSignatureAlgorithm];


export function JsonWebSignatureAlgorithmFromJSON(json: any): JsonWebSignatureAlgorithm {
    return JsonWebSignatureAlgorithmFromJSONTyped(json, false);
}

export function JsonWebSignatureAlgorithmFromJSONTyped(json: any, ignoreDiscriminator: boolean): JsonWebSignatureAlgorithm {
    return json as JsonWebSignatureAlgorithm;
}

export function JsonWebSignatureAlgorithmToJSON(value?: JsonWebSignatureAlgorithm | null): any {
    return value as any;
}